web/requirements.txt
appdirs==1.4.3
defusedxml==0.5.0
Django==1.11.1
django-allauth==0.31.0
django-debug-toolbar==1.7
django-extensions==1.7.8
django-extra-fields==0.9
django-filter==1.0.2
django-rest-auth==0.9.1
django-unused-media==0.1.6
django-versatileimagefield==1.7.0
djangorestframework==3.6.3
djangorestframework-bulk==0.2.1
djangorestframework-gis==0.11.2
flake8==3.3.0
Markdown==2.6.8
mccabe==0.6.1
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.7.1
pycodestyle==2.3.1
pyflakes==1.5.0
pyparsing==2.2.0
python3-openid==3.1.0
pytz==2017.2
redis==2.10.5
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
sqlparse==0.2.3
uuid==1.30
uWSGI==2.0.15
